// app/api/dtos/appointment.dto.ts
export interface AppointmentDto {
    appointment_id: number;
    user_id:number;
    teacher_id:number;
    first_name: string;
    last_name: string;
    category_name: string;
    start_time: string;
}
// export interface AppointmentDto {
//     appointment_id: number;
//     teacher_id: number;
//     user_id: number;
//     lesson_id: number;
//     start_time: string;
//     end_time: string;
// }